CCBUG: 168501
Hi there ; ), thanks for your feedback Karl ; )

1) About the use of "Add" and "Insert" buttons:
They have diferent functions. The first, add the main color to end of the palette and the other, insert the main color in the position of the current selected item of the palette. 

So, I changed the icons of Add and Insert, I renamed the button Add to  "Append" and moreover I will put a tooltip in the Append and Insert buttons ... 

However, I don't think that remove the Insert button is a good idea, many users would want put a color in a specific position inside the palette, unfortunately I don't have here the Basket app, but I think that with this changes the issue is solved.

2) About add colors from the tetradic scheme:
This funcion works ok, however I made some improvements, now each color of any scheme (complement or tetradic) has a button with the "plus symbol" to append the color to the palette.

3) About the brief view or palette issue:
I made all the changes that you requested, I think your point was right.

To see the changes of the points 1 and 2 please check this link (also attached to the bugtracker)

http://img48.imageshack.us/img48/1046/fixedbugsyy3.png

Regards
Percy

svn path=/trunk/extragear/graphics/kcoloredit/; revision=848092